Extract reifiers/has_many_through
